Add support for larger screens
Along with that:
- Move to different Scaffold for HomeScreen and EditScreen
   as before e3ce76844a6089932879506db63d9f518c9e483c
- Move to floating toolbar from BottomAppBar

+++ app/src/main/java/com/sarangem/zenwell/Utils.kt
@@ -38,8 +38,8 @@ fun getAmPm(num: Int): String {
 @OptIn(ExperimentalMaterial3Api::class)
 fun convertToTimePickerState(timeInMinutes: Int): TimePickerState {
     return TimePickerState(
-        initialHour = (timeInMinutes / 60).toInt(),
-        initialMinute = (timeInMinutes % 60).toInt(),
+        initialHour = (timeInMinutes / 60),
+        initialMinute = (timeInMinutes % 60),
         is24Hour = false
     )
 }
